This header will allow for the *transparent* inclusion of various system header files based upon compiler defines.
This will let us keep the #include lines in the source file nice and neat,
and abstract all the "rules" for determining what to include into one file.

** NOTE **
Not all system includes should go into this file because of conflicts.
An example is that <curses.h> and <stdarg.h> conflict if included together
in FreeBSD.  (Luckily, they don't need to be included together at this point.)