screen.c: always use "%s"-style format for printf()-style functions
`ncuses-6.3` added printf-style function attributes and now makes
it easier to catch cases when user input is used in palce of format
string when built with CFLAGS=-Werror=format-security:

    screen.c:430:17: error: format not a string literal and no format arguments [-Werror=format-security]
      430 |                 wprintw(win, _(qSdbrdonly));
          |                 ^~~~~~~

Let's wrap all the missing places with "%s" format. It should be fine
for this case as format string don't contain any format arguments.